requirements.txt
torch==2.7.0+cu121
torchvision==0.22.0+cu121
torchaudio==2.7.0+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

joblib
yt-dlp
gdown
xmltodict
pyannote.metrics
pyannote.core

einops==0.8.0
transformers==4.43.2
huggingface-hub
sentencepiece==0.1.99
wandb

encodec
ns3_codec==0.2.2
descript-audio-codec==1.0.0